-------This line and the following will be ignored --------------

modified:
  openlp/core/lib/renderer.py
  openlp/core/lib/themexmlhandler.py
  openlp/core/ui/amendthemeform.py
  openlp/core/ui/servicemanager.py
  openlp/core/ui/thememanager.py
  openlp/core/ui/themestab.py
  openlp/plugins/bibles/forms/bibleimportdialog.py
  resources/forms/bibleimportdialog.ui
pending merges:
  Tim Bentley 2009-08-07 finish Theme handling corrections 
    Tim Bentley 2009-08-07 Default handling allows editing and blocks deletes
    Tim Bentley 2009-08-07 New bible import form
    Tim Bentley 2009-08-06 [merge] Head
    Tim Bentley 2009-08-06 Standardize row hight in renderer

+++ openlp/core/ui/thememanager.py
@@ -28,7 +28,7 @@ from PyQt4 import QtCore, QtGui
 
 from openlp.core.ui import AmendThemeForm, ServiceManager
 from openlp.core.theme import Theme
-from openlp.core.lib import Event, EventType, EventManager, OpenLPToolbar, ThemeXML, Renderer,  translate,  file_to_xml,  buildIcon
+from openlp.core.lib import PluginConfig, Event, EventType, EventManager, OpenLPToolbar, ThemeXML, Renderer,  translate,  file_to_xml,  buildIcon
 from openlp.core.utils import ConfigHelper
 
 class ThemeManager(QtGui.QWidget):
@@ -68,10 +68,33 @@ class ThemeManager(QtGui.QWidget):
         self.ThemeListWidget.setAlternatingRowColors(True)
         self.ThemeListWidget.setIconSize(QtCore.QSize(88,50))
         self.Layout.addWidget(self.ThemeListWidget)
+        #Signals
+        QtCore.QObject.connect(self.ThemeListWidget,
+           QtCore.SIGNAL(u'doubleClicked(QModelIndex)'), self.changeGlobal)
+        #Variables
         self.themelist = []
         self.path = os.path.join(ConfigHelper.get_data_path(), u'themes')
         self.checkThemesExists(self.path)
         self.amendThemeForm.path = self.path
+        # Last little bits of setting up
+        self.config = PluginConfig(u'themes')
+        self.servicePath = self.config.get_data_path()
+        self.global_theme = unicode(self.config.get_config(u'theme global theme', u''))
+
+    def changeGlobal(self, index):
+        for count in range (0,  self.ThemeListWidget.count()):
+            item = self.ThemeListWidget.item(count)
+            oldName =  item.text()
+            #reset the old name
+            if oldName != unicode(item.data(QtCore.Qt.UserRole).toString()):
+                self.ThemeListWidget.item(count).setText(unicode(item.data(QtCore.Qt.UserRole).toString()))
+            #Set the new name
+            if count  == index.row():
+                self.global_theme = unicode(self.ThemeListWidget.item(count).text())
+                name = (u'(%s):%s' % (translate(u'ThemeManager', u'default'), self.global_theme))
+                self.ThemeListWidget.item(count).setText(name)
+                self.config.set_config(u'theme global theme', self.global_theme)
+                self.push_themes()
 
     def onAddTheme(self):
         self.amendThemeForm.loadTheme(None)
@@ -80,30 +103,38 @@ class ThemeManager(QtGui.QWidget):
     def onEditTheme(self):
         item = self.ThemeListWidget.currentItem()
         if item is not None:
-            self.amendThemeForm.loadTheme(unicode(item.text()))
+            self.amendThemeForm.loadTheme(unicode(item.data(QtCore.Qt.UserRole).toString()))
             self.amendThemeForm.exec_()
 
     def onDeleteTheme(self):
+        self.global_theme = unicode(self.config.get_config(u'theme global theme', u''))
         item = self.ThemeListWidget.currentItem()
         if item is not None:
             theme = unicode(item.text())
-            th = theme +  u'.png'
-            row = self.ThemeListWidget.row(item)
-            self.ThemeListWidget.takeItem(row)
-            try:
-                os.remove(os.path.join(self.path, th))
-            except:
-                #if not present do not worry
-                pass
-            try:
-                shutil.rmtree(os.path.join(self.path, theme))
-            except:
-                #if not present do not worry
-                pass
-            #As we do not reload the themes push out the change
-            self.parent.EventManager.post_event(Event(EventType.ThemeListChanged))
-            self.parent.ServiceManagerContents.updateThemeList(self.getThemes())
-            self.parent.settingsForm.ThemesTab.updateThemeList(self.getThemes())
+            # should be the same unless default
+            if theme != unicode(item.data(QtCore.Qt.UserRole).toString()):
+                QtGui.QMessageBox.critical(self,
+                    translate(u'ThemeManager', u'Error'),
+                    translate(u'ThemeManager', u'You are unable to delete the default theme!'),
+                    QtGui.QMessageBox.StandardButtons(QtGui.QMessageBox.Ok))
+            else:
+                self.themelist.remove(theme)
+                th = theme +  u'.png'
+                row = self.ThemeListWidget.row(item)
+                self.ThemeListWidget.takeItem(row)
+                try:
+                    os.remove(os.path.join(self.path, th))
+                except:
+                    #if not present do not worry
+                    pass
+                try:
+                    shutil.rmtree(os.path.join(self.path, theme))
+                except:
+                    #if not present do not worry
+                    pass
+                #As we do not reload the themes push out the change
+                #Reaload the list as the internal lists and events need to be triggered
+                self.push_themes()
 
     def onExportTheme(self):
         pass
@@ -136,11 +167,18 @@ class ThemeManager(QtGui.QWidget):
                     if os.path.exists(theme):
                         (path, filename) = os.path.split(unicode(file))
                         textName = os.path.splitext(name)[0]
-                        item_name = QtGui.QListWidgetItem(textName)
+                        if textName == self.global_theme:
+                            name = (u'(%s):%s' % (translate(u'ThemeManager', u'default'), textName))
+                        else:
+                            name = textName
+                        item_name = QtGui.QListWidgetItem(name)
                         item_name.setIcon(buildIcon(theme))
                         item_name.setData(QtCore.Qt.UserRole, QtCore.QVariant(textName))
                         self.ThemeListWidget.addItem(item_name)
                         self.themelist.append(textName)
+        self.push_themes()
+
+    def push_themes(self):
         self.parent.EventManager.post_event(Event(EventType.ThemeListChanged))
         self.parent.ServiceManagerContents.updateThemeList(self.getThemes())
         self.parent.settingsForm.ThemesTab.updateThemeList(self.getThemes())
@@ -153,6 +191,7 @@ class ThemeManager(QtGui.QWidget):
         xml_file = os.path.join(self.path, unicode(themename), unicode(themename) + u'.xml')
         try:
             xml = file_to_xml(xml_file)
+            #print xml
         except:
             newtheme = ThemeXML()
             newtheme.new_document(u'New Theme')
@@ -163,7 +202,9 @@ class ThemeManager(QtGui.QWidget):
                 unicode(0), unicode(0), unicode(0))
             xml = newtheme.extract_xml()
         theme = ThemeXML()
+        #print theme
         theme.parse(xml)
+        #print "A ", theme
         theme.extend_image_filename(self.path)
         return theme
 
@@ -211,6 +252,9 @@ class ThemeManager(QtGui.QWidget):
         self.generateAndSaveImage(dir, themename, filexml)
 
     def checkVersion1(self, xmlfile):
+        """
+        Am I a version 1 theme
+        """
         log.debug(u'checkVersion1 ')
         theme = xmlfile
         tree = ElementTree(element=XML(theme)).getroot()
@@ -220,6 +264,11 @@ class ThemeManager(QtGui.QWidget):
             return True
 
     def migrateVersion122(self, filename, fullpath, xml_data):
+        """
+        Called by convert the xml data from version 1 format
+        to the current format.
+        New fields are defaulted but the new theme is useable
+        """
         log.debug(u'migrateVersion122 %s %s', filename, fullpath)
         theme = Theme(xml_data)
         newtheme = ThemeXML()
